+++ packages/respect-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @redocly/respect-core
 
+## 2.11.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.11.0.
+
 ## 2.10.0
 
 ### Patch Changes